wifi: add controls for noscan flag
OpenWrt contains a patch for hostapd which allows to override the channel-bandwidth decission
to always use 40MHz regardless of overlapping networks.
Though it does not conform with IEEE 802.11n-2009, this became the de-facto standard behaviour
of most off-the-shelf wifi routers.

Adding this option to LuCI makes it more accessible to users, which probably not everybody will
agree with to be a good idea. However, I didn't yet come across a single commercial product which
actually complies with 802.11n-2009 in that regard.

Signed-off-by: Daniel Golle <dgolle@allnet.de>