luci-statistics: don't plot data points where there are no data.
Converting NaN to zero for the sake of the cumulative stack in non-overlay
mode is all very well. But we shouldn't then use that for displaying the
graph. Make another variable %s_plot which *preserves* NaN from the input
data, and then we don't get bogus data points drawn. Especially at the end
of the graph this was problematic.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>