modules/base: ltn12: source.file() should terminate when an empty chunk is read
The read method of nixio's file and socket objects both return an empty string
when they reach EOF, not nil, causing the consumer to loop endlessly as
source.file() never terminates.

As there is no other situation in which an empty chunk is read, just change it
to nil to terminate the consumer's loop.

Signed-off-by: Matthias Schiffer <mschiffer@universe-factory.net>