luci-app-vnstat: Fix blank graphs for iface names with underscores
The regex used to process the iface query string argument in
vnstat.htm was stripping underscores, which caused the graph not to
display for interfaces with names containing underscores.  This
patch adds the underscore to that regex so that the interface name
will be correct and the graph will be displayed.

Signed-off-by: Michael Marley <michael@michaelmarley.com>