for-15.05 luci-app-firewall: use maxlength datatype for zone name validation
Backport from trunk the zonename validation using
a compound datatype "and(uciname,maxlength(14))"

fw3 sets the maximum length of the zone name to 14 and
ignores zone definitions with too long names.
Add a simple validation to ensure that the new zone name is short enough.

Backport of 34e875b3 and edc58332

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>