Don't force PolarSSL as SSL library and remove key generator
The uhttpd-mod-tls module that is required for handling HTTPS can be build against several cryptographic libraries. Instead of forcing LuCI to use the PolarSSL library, depend on uhttpd-mod-tls.

Strictly speaking it is not necessary to depend on px5g since that's only needed to generate a key. If a key is already available, there is no need for px5g.