Fix check for musl library
Checking whether or not the second argument is actually used by 'nslookup' should not use a valid IP address. When compiled with the 'musl' library, this argument is ignored (no matter what it is). But when it *is* used, using '0' will result in Busybox nslookup to attempt to query the IP_ANYADDR '0.0.0.0'. This may typically take up to 15 seconds to timeout. It is better to provide an illegal argument, which will immediately return with an error.

Signed-off-by: Arjen M de Korte \<build+openwrt@de-korte.org\>