luci-base: Include Openwrt build revision in displayed version string
The version string displayed in the footer and the overview page contains
both LuCI and Openwrt versions. But due to differences in the contents
of /etc/openwrt_release the exact Openwrt revision r..... may be missing:
* In trunk that revision is included in the DISTRIB_DESCRIPTION field that
  is displayed in Luci as the Openwrt version.
* In release branches, that field contains the release number like 15.05.

Revision number is always present in DISTRIB_REVISION field, so add the
revision from that field in case the revision is not already present
in the string.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>
(cherry picked from commit 50451c4ac53c853cbf2837e7bfbbc1268a440180)