luci-app-pbx(-voicemail): mark BROKEN
Mark the luci-app-pbx(-voicemail) packages BROKEN as trying
to install them now produces dependency errors.

The pbx packages depend on asterisk 1.8 that is EOL upstream
and has been moved from the telephony feed to the abandoned feed.

If LuCI pbx packages are still needed, they should be refreshed
to depend on a current asterisk version from the telephony feed.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>