wifi_add: allow multi SSID by default if supported.
Instead of defaulting to replace, even on hardware that supports multi
SSID, make the default to add.  This reduces the chance of locking
oneself out when adding a second network.  If you intended to replace,
ending up with two networks then deleting one is easier than recovering
when the old network configuration is gone.

Signed-off-by: Karl Palsson <karlp@etactica.com>