luci-mod-admin-full: Allow to generate fstab when not present
This fixes the situation that one cannot use the mount points GUI
until one has use the cli to create an fstab config file.  Instead
we always show system mount points, and if block-mount is present
we allow to generate and fstab and if an fstab is present we show
it.

Signed-off-by: Daniel Dickinson <lede@cshore.thecshore.com>