luci-mod-admin-full: add config possibility to hide led config page
In some situation it is neccassery to hide the configuration possibility
for some pages. In this example the led page.

Under the config page /etc/config/luci there is a new config area with
the name "show". The page will always shown until you add the name
"AdminSystemLeds" with value "0" to the config.

For now the page "AdminSystemLeds" is configurable. See commit.

Signed-off-by: Florian Eckert <Eckert.Florian@googlemail.com>