luci-app-fwknopd: fix for qrencode >= v3.4.0.
Starting with qrencode v3.4.0, the XML prolog is prepended to the SVG
output which causes an XML parse error.  To address this, strip any XML
prolog from the generated output.

Also, as a precaution, explicitly set the qrencode output type to SVG to
to protect against a change in defaults.