luci-proto-wireguard: enable addressing for tunnel interfaces
Enable static addresses on WireGuard tunnel interfaces without requiring
an static address interface.

This removes the requirement to use a static address interface on top of a
WireGuard tunnel interface in the majority of cases. In the past, users have
been confused by the current approach and asked for a simpler way to configure
WireGuard interfaces.

Signed-off-by: Dan Luedtke <mail@danrl.com>