luci-base: fix langage missmatch on auto setting
missmatch laungage when auto settings.
It is a way to help things that are not perfect matches, but I would like to hear opinions.

example,
browser setting is ja-JP, I want you to selecting ja on auto setting.
In many cases, browser setting is ja-JP or ja-XX, not ja. Sadly.
fix to help these.

Signed-off-by: YuheiOKAWA <tochiro.srchack@gmail.com>