luci-app-attendedsysupgrade: use common
use attendedsysupgrade-common package and rpc-sys instead of
rpcd-mod-attendedsysupgrade

luci-app-attendedsysupgrade: editable server url

the server url can now be changed. The solution is kept as simple as
possible to not introduce any lua code.

Small cleanup, but needs a refactoring.

Add acl rights to edit the attendedsysupgrade file

Add cgi-io as a dependencie to upload image to router

Signed-off-by: Paul Spooren <paul@spooren.de>