luci-mod-admin-full: change check support_reset test for factory reset
Redmine-patch-id: 1842
The check "support_reset" only covers /proc/mtd partitions. If we have
this the commands checks for names like ubi or rootfs_data. If this is found
the system provieds a factory reset.
But on x86 the situation is difference. We have no /proc/mtd partitions
because this system do not use a bare metall flash.

To solve this issue check if we have an overlay and if so we could do a
factory reset. This could be applied for system which uses bare metal flash
and system which uses FTL or harddisks.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>