luci-app-storage: created
This luci app provides a basic interface to prepare a single
drive for storage without having to use SSH command line tools.

It lets the user choose the drive and filesystem, it will then
wipe the drive, create a new GPT partition spanning the whole drive,
and format it with the chosen filesystem.

It does not show in the interface (therefore the user cannot format)
drives/partitions used by the system, mtd partitions, and mdadm RAID.

Since generating the html templates in /view/storage to allow enough
drive and filesystem choice would be extremely tedious, I provide a
"developer helper" script that automates the task, which should be
used by anyone that wants to add a new filesystem or more drive names.

Signed-off-by: Alberto Bursi <alberto.bursi@outlook.it>