luci-app-ddns: remove block-loading function
Load the massive ddns tools function adds 10s delay to luci interface, this cause every XHR requst that refresh every 5 second to fail. The "ddns" variable is not even used in index function, so it's just a waste of resource and cause lots of problem. By removing itall the delay problem are solved.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>