luci-mod-admin-full: fix broken applyreboot page
Currently applyreboot page lost the style and result in an empty page with some text. With this the page renders correctly. Also we fix problems when server address is broken by taking the ipaddr from the browser directly and we make the flash message more readable. We move the style.css file from luci-theme-material to cascade.css and we change the header css to the default construction.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>