luci-app-firewall: Ease creation of multiple day traffic rules
Creating a traffic rule that spans multiple days currently requires
creating three rules, unless the rule only covers complete days. You
need to create one rule for the first day, one rule for the complete
days, and then one rule for the final day. In addition to making
creating multi day-rules a bit of as hassle, performance suffers when
many multi day-rules are present on a system.

The --datestart/--datestop arguments to iptables, which is what
Start/Stop Date maps to, supports specifying a time of day. This PR
extends Start/Stop Date to also accept time. The new format of the
fields are yyyy-mm-ddThh:mm:ss. As before, Start/Stop Time will override
the time specified in Start/Stop Date.

I did not find a nice way to make the time-part of Start/Stop Data
optional. I don't believe this matters to much. If a user edits an
existing rule, he or she will get a validation error when the rule is
saved. However, all that has to be done to make the date(s) pass
validation (and keep the old behavior) is to add "T00:00:00" to the date
field(s).

The translations are also updated, with the exception of Russian and
Ukrainian. My knowledge of the cyrillic alphabet is a bit limit, to say
the least.

Signed-off-by: Kristian Evensen <kristian.evensen@gmail.com>