luci-app-firewall: Ease creation of multiple day traffic rules
Creating a traffic rule that spans multiple days currently requires
creating three rules, unless the rule only covers complete days. You
need to create one rule for the first day, one rule for the complete
days, and then one rule for the final day. In addition to making
creating multi day-rules a bit of as hassle, performance suffers when
many multi day-rules are present on a system.

The --datestart/--datestop arguments to iptables, which is what
Start/Stop Date maps to, supports specifying a time of day. This PR
extends Start/Stop Date to also accept an optional time. The format of
the fields are now either yyyy-mm-dd or yyyy-mm-ddThh:mm:ss. As before,
Start/Stop Time will override the time specified in Start/Stop Date.

v1->v2:
* Update commit message
* Remove markup from translation strings.
* Extend existing datatype instead of adding a new type. I have added a
flag "allow_time" to "dateyyyymmdd" that, when set, enables users to
pass a time in addition to date.
* The time-part of "dateyyyymmdd" is optional. I decided to check if
time is passed or not by checking the length of val. By checking length,
we save a potentially redundant call to match.
* Added Russian message strings (thanks namedun).

Signed-off-by: Kristian Evensen <kristian.evensen@gmail.com>