Bootstrap at Night : "luci-theme-nightstrap" …
I've created a real Nightmode for OpenWrt based on the Bootstrap-Theme. This Commit contains a tweaked "cascade.css" and "mobile.css" with the new Colorization and a few Fixes, so it looks much better than the original Bootstrap-Theme. The "header.htm" and "footer.htm" are unchanged …

Hopefully this Packages will be included in a further Release of OpenWrt.