luci-base: limit new pkg version on release branches
This commits prevents new package version when no commit touches package
files. It is limited to branches and does not change master behavior.

Adds GIT_COMMIT to be used as reference for package version. For master
branch, GIT_COMMIT will be HEAD (as before). For branches, GIT_COMMIT
will be the most recent common ancestor between the branch and master.

Adds GIT_COMMIT_COUNT as the number of commits inside each package
directory since GIT_COMMIT. For master, it should be always 1.

Adds LUCI_RELEASE as the global LuCI release control. It should be
bumped to generate a new release for all packages. For master, it should
be always 1.

PKG_RELEASE will be the sum of LUCI_RELEASE and GIT_COMMIT_COUNT.

svn logic was removed and git command simplified.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>