luci-app-attendedsysupgrade: Fix "Edit packages" error
Function edit_packages() expects data.packages to be an array of
packages and bombs out if it isn't. But request_json.packages is not
always defined (e.g. if the release did not change but only some extra
packages need to be upgraded).

Set data.packages is set to the list of currently installed
packages in this case.

Signed-off-by: Richard Kunze <richard.kunze@web.de>