luci-app-attendedsysupgrade: Fix server error in "upgrade-request"
"upgrade-request" bombs out with an HTTP 500 error if "version" is
missing from the request, but data.latest_version is only set if an upgrade
to a new version is available, not if only some packages need to be
upgraded. In this case, use the currently installed version for the
upgrade request.

Signed-off-by: Richard Kunze <richard.kunze@web.de>