luci-mod-system: add application led trigger possibility
Currently we only support native kernel LED triggers.
In some situations it is useful to have LED triggers for system events.
Let's call it application triggers.

An example would be the status of all connections on the mwan3.

There are the following possible states:
* If all mwan3 connections are connected, then the alarm LED is off
  (kernel trigger -> none).
* If all mwan3 connections are not connected, then the alarm LED is on
  (kernel trigger -> default-on).
* If one of the mwan3 connections is not connected (backup), the alarm
  LED should flash (kernel trigger -> timer).

The evaluation is done by a script in the directory
"/etc/hotplug.d/iface"

Currently it is not possible to select this type of trigger. This change
makes this possible like with the Proto-Handlers in LuCI.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>