luci-app-statistics: update misleading load view
Fixes: #3301

During my tests and developing with the collectd and so
luci-app-statistics I noticed that the values displayed in the Y-axis
are misleading, because probably the individual values are added
together in each case. So the view is not corrected.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>

(cherry picked from commit b50974a20c9dac0d6b8b479d2e6b54b3c5703ab6)
Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>