luci-base: add queryable WEP WiFi feature
This commit adds WEP as a queryable WiFi feature.

Support for the deprecated WEP encryption is not
compiled-in to hostapd or wpa_supplicant by default
anymore.

Allow LuCi to query the availability of WEP to remove
it from the list of avialable encryption methods in case
hostapd / wpa_supplicant are compiled without it.

Signed-off-by: David Bauer <mail@david-bauer.net>