New Restrictive, Moderated or Disabled options for YouTube SafeSearch
New options to choose from for YouTube SafeSearch: Restrictive, Moderated and Disabled. This will help with some issues that might be encountered even when selecting Moderate options, YouTube tends to be too restrictive with some content even under this filter, a good example are some music. It becomes an annoyance, so being able to disable SafeSearch only for YouTube is very handy.

Signed-off-by: Alberto Martínez Álvarez <amteza@gmail.com>
Signed-off-by: Alberto Martínez Álvarez <amteza@gmail.com>