luci-app-dawn: index SSID section ids
HTML requires the "id" attribute on an element to be unique. In case of
networks with multiple SSIDs, the code would output the same id value
multiple times, which is technically invalid.

Since the SSID can contain characters that would be invalid in an id
attribute, use a running index to create unique values.

Signed-off-by: Sander Vanheule <sander@svanheule.net>