luci-app-dawn: don't show bare template string
It might occur that there is no signal strength reported in the network
overview. In that case clientsvals.signal would be nil, and the template
string "%d" would not be processed.

Check if clientvals.signal is different from nil before trying to format
the value. Otherwise return "?" as a placeholder.

Signed-off-by: Sander Vanheule <sander@svanheule.net>