luci-mod-network: add ppsk option (support for private psk)
Include support in luci to enable the Private PSK.
When Private PSK is enabled, clients can use a default password (common),
or have their own private password that is associate with the client MAC.

The password is retrieved from RADIUS server, asking for the client MAC,
and when such client MAC is not defined, RADIUS should return a default
password.

RADIUS can also return other parameters like VLANs, which can be used to
put clients dinamically in different vlans based on their MAC, or the
default configuration.

Private PSK is not compatible yet with SAE encryption, therefore cannot
be used yet with WPA3 or WPA3-mixed.

It implements the UI for the PPSK already in master: https://github.com/openwrt/openwrt/pull/3509

Signed-off-by: Manuel Giganto <mgigantoregistros@gmail.com>