luci-app-firewall: map proto '*' and 'any' to all on rule config
Before the change, the options '*' and 'any' in the drop down were not
recognized as valid options, when loaded from the uci. With this change,
the options '*' and 'any' are mapped to 'all' and saved as such. This
change is especially important if the proto option is changed manually
to '*' or 'any' in shell and then further configured via LuCI.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 972096bf39c932ad007fb431c81a2f5fab64649f)