luci-app-sqm: only reload view once sqm enable/start completed
Chain the exec calls to enable and start the SQM service and only reload
the view once these operations are complete and return the entire promise
chain from the handler function to ensure that the busy indicator remains
active throughout the entire duration of the operation.

Also replace exec_direct() with exec() as there is no need to bypass ubus
for the init script calls.

Fixes: e76d9cc767 ("luci-app-sqm: fix JS error if service is disabled")
Signed-off-by: Jo-Philipp Wich <jo@mein.io>