luci-app-acl: unset read and write before acl set
If the setting in the view is set to `denied`, only the read list option
is deleted. This is not correct. The write list option must also be
deleted. To ensure that the correct configuration is saved, the write
and read list options are always deleted beforehand and then rewritten.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit f4e8fcb6b53780bd2d126b22a89c379cd7c071d7)