luci.mk: respect `version` file in package folders
LuCI generates package versions either by the last git commit or the
content of the `SOURCE_DATE_EPOCH` env variable. This commit allows to
use the content of a `version` file if present.

This should allow a better reproducibility when creating releases. The
`version` file per folder can be created by the (extended) maintenance
script which is used to generate OpenWrt releases.

Signed-off-by: Paul Spooren <mail@aparcar.org>