luci-mod-network: diagnostics: remove arp-scan install placeholder
We do not want to advertise functionality that is not installed as it may
have been left out intentionally. Referring to "Missing" functionality
is confusing in this case. It is also not comparable to missing protocol
extension hints as those are backed by actual user configuration.

Also replace the freetext input with a dropdown of known network devices
as the hardcoded `br-lan` is not universally applicable to all setups.

In general we want to avoid hardcoding defaults which might deviate from
the system state.

Finaly drop the superfluous width annotations and add a missing space
to the RPC command declaration.

Fixes: 22b7442011 ("luci-mod-admin-full: add arp-scan to network diagnostic")
Signed-off-by: Jo-Philipp Wich <jo@mein.io>