luci-mod-network: wireless.js Extended cell_density
Add LuCI support for the enhanced cell_density option that manages
STA association requests or probe requests to gate to appropriate RSSI
thresholds for the cell densities.

New cell_density options are used to support additional
threshold RSSIs at very high cell densities.

Cell Density

0 - Disabled (Default)
Setting cell_density to 0 does not configure a threshold RSSI.
Setting cell_density to 0 does not configure data rates.

1 - Low (Normal)
Setting cell_density to 1 configures a threshold RSSI of -80 dBm.
Setting cell_density to 1 configures the basic/mandatory rates to 6, 12
and 24 Mb/s OFDM rates where legacy_rates is 0. Supported rates lower
than the minimum basic/mandatory rate are not offered.
Setting cell_density to 1 configures the basic/mandatory rates to the 5.5
and 11 Mb/s DSSS rates where legacy_rates is 1. Supported rates lower
than the minimum basic/mandatory rate are not offered.

2 - Medium
Setting cell_density to 2 configures a threshold RSSI of -75 dBm.
Setting the cell_density to 2 configures the basic/mandatory rates to the
12 and 24 Mb/s OFDM rates where legacy_rates is 0. Supported rates lower
than the minimum basic/mandatory rate are not offered.
Setting the cell_density to 2 configures the basic/mandatory rates to the
11 Mb/s DSSS rate where legacy_rates is 1. Supported rates lower than the
minimum basic/mandatory rate are not offered.

3 - High
Setting cell_density to 3 configures a threshold RSSI of -70 dBm.
Setting the cell_density to 3 configures the basic/mandatory rates to the
24 Mb/s OFDM rate where legacy_rates is 0. Supported rates lower than the
minimum basic/mandatory rate are not offered.

4 - High
Setting cell_density to 4 configures an RSSI threshold of -67 dBm.
Setting the cell_density to 4 configures the basic/mandatory rates to the
24 Mb/s OFDM rate where legacy_rates is 0. Supported rates lower than the
minimum basic/mandatory rate are not offered.

5 - High
Setting cell_density to 5 configures an RSSI threshold of -65 dBm.
Setting the cell_density to 5 configures the basic/mandatory rates to the
24 Mb/s OFDM rate where legacy_rates is 0. Supported rates lower than the
minimum basic/mandatory rate are not offered.

6 - Very High
Setting cell_density to 6 configures a threshold RSSI of -62 dBm.
Setting the cell_density to 6 configures the basic/mandatory rates to the
24 Mb/s OFDM rate where legacy_rates is 0. Supported rates lower than the
minimum basic/mandatory rate are not offered.

7 - Very High
Setting cell_density to 7 configures a threshold RSSI of -60 dBm.
Setting the cell_density to 7 configures the basic/mandatory rates to the
24 Mb/s OFDM rate where legacy_rates is 0. Supported rates lower than the
minimum basic/mandatory rate are not offered.

Signed-off-by: Nick Lowe <nick.lowe@gmail.com>