luci-mod-system: Add field to allow user-customizable branding
Add a field in the "Language and Style" tab of System to allow users to override
the title/branding of Luci (even emoji).  If left empty, behavior reverts to using the hostname.

Also tweak the default themes to have the hostname displayed in the footer so as not to
completely remove it from view.

Signed-off-by: Tim Thorpe <tim@tfthorpe.net>