luci-material-theme: align td text width 33%
The CSS applies a different padding for td elements with 33% width. This
misaligns the text from this td to the others td.

This change modifies the CSS to apply the same padding to all of the
sibling td.

Signed-off-by: Miguel Angel Mulero Martinez <migmul@gmail.com>
(cherry picked from commit 3333fd6695cbd2e7a6055f01542242a3b7c111ec)