Updates proto, mssfix, user & group conds and defs
This parameter is mandatory for other parameters of server or client modes (topology, mssfix...) but the condition was to restrict it to client mode only. This parameter is also moved to the more appropriate "networking" section.
The default value of mssfix is set to the value documented by openvpn.org.
The user and group params are uncommented to be available in "service" section with secured default values.