luci-app-upnp: Fix rpc handler for nftables
Rewrite great portions of the rpcd handler script to utilize the
newer nftables implementation for firewall4-using versions. This
is required because miniupnpd-nftables creates tables that are
incompatible with iptables-nft. There may be some race conditions
with deleting the lease entries, since it opens the file to
locate the line number to delete. There could be a better way to
do this.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>