luci-compat: fix common cbi-page-actions position
With legacy applications, the common page actions "Save and Apply" are
not displayed correctly, so that they are not aligned on the right as
with javascript applications. To fix the alignment, the control-group
css class is removed.

This was determined with the application 'luci-app-nft-qos'
with the theme 'luci-theme-bootstrap'.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit a70c5705e476cf1325edceab29c15148c289f280)