luci-app-mwan3: fix i18n plural warning
Updating applications/luci-app-mwan3/po/templates/mwan3.pot ...
xgettext: warning: msgid '%d second' is used without plural and with
plural.
  standard input:175: Here is the occurrence without plural.
  standard input:170: Here is the occurrence with plural.
Workaround: If the msgid is a sentence, change the wording of the
sentence; otherwise, use contexts for disambiguation.
xgettext: warning: msgid '%d second' is used without plural and with
plural.
  standard input:191: Here is the occurrence without plural.
  standard input:170: Here is the occurrence with plural.
Workaround: If the msgid is a sentence, change the wording of the
sentence; otherwise, use contexts for disambiguation.
xgettext: warning: msgid '%d second' is used without plural and with
plural.
  standard input:213: Here is the occurrence without plural.
  standard input:170: Here is the occurrence with plural.
Workaround: If the msgid is a sentence, change the wording of the
sentence; otherwise, use contexts for disambiguation.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>