luci-mod-network: rename button reset to delete on device page
If a device is created that did not exist before, the description
deconfigure is not correct. When the 'Deconfigure' button is pressed,
the interface is deleted and removed from the list. This is not what one
would expect. Therefore, it is more appropriate to rename the button to
delete. This is rather the function that is executed when the button is
pressed. Devices that cannot be created dynamically cannot be
deconfigured and thus deleted.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>