luci-light: rework and reenable
The `luci` collection contains package management and attended
sysupgrades support, which are not that useful if not even undesired for
self-compiled images.

Rework the `luci-light` collection to exclude the two above mentioned
features, and make `luci` instead depend on the light collection in
additon to those features.

The `luci-ssl` and `luci-ssl-openssl` collections then only need
to depend on `luci-light`.

We now have three variants with won't pull in `luci-app-opkg` or
`luci-app-attendedsysupgrade`, making everybody happy.

Signed-off-by: Andre Heider <a.heider@gmail.com>