luci-mod-network: remove uppercase for interface name on overview page
In the overview page, the name of the interface is converted to
uppercase. However, this is not the name in the configuration. From my
point of view, this makes no sense. The name displayed should correspond
exactly to the name in the configuration.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 5ffaf478424897e856c290b1bc32d6c3221ee04c)