luci-app-statistics: set shortest period data/graph to 2 hours
Change the shortest defined statistics period from 1 hour to 2 hours.

In practice, this only changes the graph for the shortest period
to show 2 hours of data instead of 1 hour.

The underlying database is not changed:
there are currently 288 data items for each period,
so with the 30 seconds default step, the shortest data series
contains 288 x 0.5 min = 144 min > 2 hours of data.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>