luci-base: fix button handling for named sections
If more than one named section is added to the page, it is currently the
case, that the first button is always switched on or off during input
validation of the uci seciton name.
This is because the current usage of the 'document.querySelector' function
uses the CSS classes. However, this is not unique if several named sections
are used. By setting a unique ID for the button, the function
'document.querySelector' can now use the ID and thus the correct button
is switched on or off during validation of uci section name.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>