luci-base: fix sort order for size columns
Currently elements of package size, packet size, etc. columns are
interpreted as text strings, so we have "800 B" > "7.54 KiB" > "5.11 MiB".
To fix this, a new regexp is added to UITable.deriveSortKey.

It is still not a perfect fix for System->Software->Installed, since
pagination is done based on package names, so within each page the packages
are sorted correctly, but the smallest (resp., largest) package is not
necessarily on the first (resp., last) page.

Closes: #5994 (see also #6120)

Signed-off-by: Rani Hod <rani.hod@gmail.com>